pyproject.toml
[tool.poetry]
name = "tinygs-api-server"
version = "0.1.0"
description = ""
authors = ["jarp0l"]
readme = "README.md"
packages = [{include = "tinygs_api_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
pymongo = "^4.3.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.6.0"
python-decouple = "^3.7"
pydantic = {extras = ["email"], version = "^1.10.4"}
uvicorn = {extras = ["standard"], version = "^0.20.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
